Fixed AppCode erros and warnings.

+++ Beezle/Classes/State/Debug/TestAnimationsState.m
@@ -143,7 +143,7 @@
 -(void) update:(ccTime)delta
 {
 	[_world loopStart];
-	[_world setDelta:(1000.0f * delta)];
+	[_world setDelta:(int)(1000.0f * delta)];
     
     [_renderSystem process];
 }
